Clean up AsyncRewriter tests, add test cases to check that output actually compiles.
