--- tests/Shaolinq.AsyncRewriter.Tests/ICommand.cs
@@ -1,10 +1,0 @@
-﻿// Copyright (c) 2007-2017 Thong Nguyen (tumtumtum@gmail.com)
-
-namespace Shaolinq.AsyncRewriter.Tests
-{
-	[RewriteAsync]
-	public interface ICommand
-	{
-		void Test();
-	}
-}
