+++ tests/Shaolinq.AsyncRewriter.Tests/RewriteTests/ICommand.cs
@@ -1,0 +1,10 @@
+﻿// Copyright (c) 2007-2017 Thong Nguyen (tumtumtum@gmail.com)
+
+namespace Shaolinq.AsyncRewriter.Tests
+{
+	[RewriteAsync]
+	public partial interface ICommand
+	{
+		void Test();
+	}
+}
