+++ tests/Shaolinq.AsyncRewriter.Tests/RewriteTests/TestExpressionBody.cs
@@ -1,0 +1,21 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Shaolinq.AsyncRewriter.Tests
+{
+	public partial class TestExpressionBody
+	{
+		[RewriteAsync]
+		public bool Test() => Foo();
+		
+
+		[RewriteAsync]
+		public bool Foo()
+		{
+			return true;	
+		}
+	}
+}
