+++ tests/Shaolinq.AsyncRewriter.Tests/RewriteTests/TestGenericSpecialisedImplementation.cs
@@ -1,0 +1,24 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Shaolinq.AsyncRewriter.Tests
+{
+    public partial class BaseClass<T>
+    {
+        [RewriteAsync]
+        public virtual void Foo<D>(T parameter, D domainObject)
+        {
+        }
+    }
+
+    public partial class DerivedClass : BaseClass<Guid>
+    {
+        [RewriteAsync]
+        public override void Foo<E>(Guid parameter, E domainObject)
+        {   
+        }
+    }
+}
